backend/requirement.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy>=1.4,<2.0
pymysql>=1.0.2
pyjwt==2.8.0
bcrypt==4.0.1
python-multipart==0.0.6
cryptography==41.0.5
asyncmy
google-generativeai>=0.3.0
grpcio>=1.60.0
protobuf>=4.25.0
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.20.0